Sistema de Login criado e funcionando perfeitamente.

+++ app/Controllers/Noticias.php
@@ -5,10 +5,12 @@
 
     class Noticias extends Controller {
         public function index() {
+
             $model = new NoticiasModel();
             $data = [
                 'title' => 'Últimas Noticias',
-                'noticias' => $model->getNoticias()
+                'noticias' => $model->getNoticias(),
+                'session' => \Config\Services::session(),
             ];
 
             echo view('templates/header',$data);
@@ -17,6 +19,8 @@
         }
 
         public function item($id = NULL){
+
+            $data['session'] = \Config\Services::session();
 
             $model = new NoticiasModel();
 
@@ -35,6 +39,11 @@
 
         public function inserir(){
 
+            $data['session'] = \Config\Services::session();
+            if(!$data['session']->get('logged_in')) {
+                return redirect('login');
+            }
+
             helper('form');
 
             $data['title'] = 'Inserir Noticias';
@@ -50,8 +59,13 @@
 
              $data = [
                 'title' => 'Editar Noticia',
-                'noticias' => $model->getNoticias($id)
+                'noticias' => $model->getNoticias($id),
+                'session' => \Config\Services::session(),
              ];
+
+             if(!$data['session']->get('logged_in')) {
+                return redirect('login');
+            }
 
               if (!empty($data['noticias]'])){
                 throw new \CodeIgniter\Exceptions\PageNotFoundException('Não é possivel 
@@ -65,6 +79,12 @@
         }
 
         public function gravar(){
+
+            $data['session'] = \Config\Services::session();
+            if(!$data['session']->get('logged_in')) {
+                return redirect('login');
+            }
+
             helper('form');
             $model = new NoticiasModel();
 
@@ -92,6 +112,12 @@
         }
 
         public function excluir($id = NULL) {
+
+            $data['session'] = \Config\Services::session();
+            if(!$data['session']->get('logged_in')) {
+                return redirect('login');
+            }
+
             $model = new NoticiasModel();
             $model->delete($id);
             return redirect('noticias');
